rename mayors to mayordata
