+++ src/main/java/at/hannibal2/skyhanni/data/MayorData.kt
@@ -1,7 +1,7 @@
 package at.hannibal2.skyhanni.data
 
-import at.hannibal2.skyhanni.data.MayorAPI.currentMayor
-import at.hannibal2.skyhanni.data.MayorAPI.foxyExtraEventPattern
+import at.hannibal2.skyhanni.data.ElectionAPI.currentMayor
+import at.hannibal2.skyhanni.data.ElectionAPI.foxyExtraEventPattern
 import at.hannibal2.skyhanni.data.Perk.Companion.toPerk
 import at.hannibal2.skyhanni.data.jsonobjects.other.MayorPerk
 import at.hannibal2.skyhanni.test.command.ErrorManager
