+++ src/main/java/at/hannibal2/skyhanni/features/gui/customscoreboard/elements/ScoreboardElementMayor.kt
@@ -1,6 +1,6 @@
 package at.hannibal2.skyhanni.features.gui.customscoreboard.elements
 
-import at.hannibal2.skyhanni.data.MayorAPI
+import at.hannibal2.skyhanni.data.ElectionAPI
 import at.hannibal2.skyhanni.features.gui.customscoreboard.CustomScoreboard.mayorConfig
 import at.hannibal2.skyhanni.features.rift.RiftAPI
 import at.hannibal2.skyhanni.utils.TimeUtils.format
@@ -9,15 +9,15 @@ import at.hannibal2.skyhanni.utils.TimeUtils.format
 // set 1s timer
 object ScoreboardElementMayor : ScoreboardElement() {
     override fun getDisplay() = buildList {
-        val currentMayorName = MayorAPI.currentMayor?.mayorName?.let { MayorAPI.mayorNameWithColorCode(it) } ?: return@buildList
+        val currentMayorName = ElectionAPI.currentMayor?.mayorName?.let { ElectionAPI.mayorNameWithColorCode(it) } ?: return@buildList
         val timeTillNextMayor = if (mayorConfig.showTimeTillNextMayor) {
-            "§7 (§e${MayorAPI.nextMayorTimestamp.timeUntil().format(maxUnits = 2)}§7)"
+            "§7 (§e${ElectionAPI.nextMayorTimestamp.timeUntil().format(maxUnits = 2)}§7)"
         } else ""
 
         add(currentMayorName + timeTillNextMayor)
 
         if (mayorConfig.showMayorPerks) {
-            MayorAPI.currentMayor?.activePerks?.forEach { perk ->
+            ElectionAPI.currentMayor?.activePerks?.forEach { perk ->
                 add(" §7- §e${perk.perkName}")
             }
         }
@@ -32,21 +32,21 @@ object ScoreboardElementMayor : ScoreboardElement() {
     override fun showIsland() = !RiftAPI.inRift()
 
     private fun addMinister() = buildList {
-        val ministerName = MayorAPI.currentMinister?.mayorName?.let { MayorAPI.mayorNameWithColorCode(it) } ?: return@buildList
+        val ministerName = ElectionAPI.currentMinister?.mayorName?.let { ElectionAPI.mayorNameWithColorCode(it) } ?: return@buildList
         add(ministerName)
 
         if (mayorConfig.showMayorPerks) {
-            MayorAPI.currentMinister?.activePerks?.forEach { perk ->
+            ElectionAPI.currentMinister?.activePerks?.forEach { perk ->
                 add(" §7- §e${perk.perkName}")
             }
         }
     }
 
     private fun addJerryMayor() = buildList {
-        val jerryExtraMayor = MayorAPI.jerryExtraMayor
+        val jerryExtraMayor = ElectionAPI.jerryExtraMayor
         val extraMayor = jerryExtraMayor.first ?: return@buildList
 
-        val extraMayorName = MayorAPI.mayorNameWithColorCode(extraMayor.mayorName)
+        val extraMayorName = ElectionAPI.mayorNameWithColorCode(extraMayor.mayorName)
         val extraTimeTillNextMayor = if (mayorConfig.showTimeTillNextMayor) {
             " §7(§6${jerryExtraMayor.second.timeUntil().format(maxUnits = 2)}§7)"
         } else ""
