+++ src/main/java/at/hannibal2/skyhanni/features/inventory/MinisterInCalendar.kt
@@ -2,7 +2,7 @@ package at.hannibal2.skyhanni.features.inventory
 
 import at.hannibal2.skyhanni.SkyHanniMod
 import at.hannibal2.skyhanni.data.Candidate
-import at.hannibal2.skyhanni.data.MayorAPI
+import at.hannibal2.skyhanni.data.ElectionAPI
 import at.hannibal2.skyhanni.events.InventoryCloseEvent
 import at.hannibal2.skyhanni.events.InventoryOpenEvent
 import at.hannibal2.skyhanni.events.render.gui.ReplaceItemEvent
@@ -41,11 +41,11 @@ object MinisterInCalendar {
     @SubscribeEvent
     fun onInventoryOpen(event: InventoryOpenEvent) {
         if (!isEnabled()) return
-        if (!MayorAPI.calendarGuiPattern.matches(InventoryUtils.openInventoryName())) return
-        val minister = MayorAPI.currentMinister ?: return
+        if (!ElectionAPI.calendarGuiPattern.matches(InventoryUtils.openInventoryName())) return
+        val minister = ElectionAPI.currentMinister ?: return
 
         val itemStack = "${minister.name}_MAYOR_MONSTER".asInternalName().getItemStack()
-        val ministerColor = MayorAPI.mayorNameToColorCode(minister.mayorName)
+        val ministerColor = ElectionAPI.mayorNameToColorCode(minister.mayorName)
 
         ministerItemStack = changeItem(ministerColor, minister, itemStack)
     }
@@ -53,7 +53,7 @@ object MinisterInCalendar {
     @SubscribeEvent
     fun onInventoryClose(event: InventoryCloseEvent) {
         if (!isEnabled()) return
-        if (!MayorAPI.calendarGuiPattern.matches(InventoryUtils.openInventoryName())) return
+        if (!ElectionAPI.calendarGuiPattern.matches(InventoryUtils.openInventoryName())) return
         ministerItemStack = null
     }
 
@@ -61,7 +61,7 @@ object MinisterInCalendar {
     fun replaceItem(event: ReplaceItemEvent) {
         if (!isEnabled()) return
         if (event.inventory !is ContainerLocalMenu || event.slot != MINISTER_SLOT) return
-        if (!MayorAPI.calendarGuiPattern.matches(InventoryUtils.openInventoryName())) return
+        if (!ElectionAPI.calendarGuiPattern.matches(InventoryUtils.openInventoryName())) return
         event.replace(ministerItemStack ?: return)
     }
